Changed Save Feature
Added a check to ensure that the date processed should be saved. Also included date in saved information to compare versions if needed. Need to fix override issue but not high priority at the moment. Also edited the save feature to include the name of the subject for easier identification.