Updated Peak Calculations
Added 1 second averages, mmvc is centered on the peak with half a second before and half a second after the peak. In cases where it is is too close to the start or end the 1 second starts at the beginning or starts 1 second prior to end. The same is done for smvc but the 1 second average starts once the peak has been achieved. The same method is used if the peak is towards the end, if it is too close to the end of the collection it is collected from the last second.